arcd/Makefile: Fixes #90, use ${CURDIR} instead of ${PWD}
PWD is not guaranteed to be passed to the Makefile
for example when a target task is executed with
sudo which sanitizes the ENV. It has also been
found as not being set with GNU Make running
on OpenBSD/amd64.

Instead we use ${CURDIR} which is set by GNU Make
itself and will work properly even with recursive
make calls.

see:
https://www.gnu.org/software/make/manual/html_node/Recursion.html

Signed-off-by: Adam Wolk <adam.wolk@tintagel.pl>